[PAM-3196] Don't close httpclient after every request to pam-search-api
HttpClient was constructed anew for every request made to pam-search-api, since
it was obtained from factory in a 'use' block in ktor request handler (which
closes it automatically). This is inefficient, the http client needs to be
reused, so we can take advantage of connection keep-alives to pam-search-api,
and so forth.

So, instead of supplying an http-client factory to the get-feed-handler, provide
an already instantiated httpclient that is kept and reused.

Also, shouldn't be necessary to wrap http-client calls in async-blocks, as
context of call is already async in ktor request handler.